stocks
dodataki

+++ Models/StockTickers.cs
@@ -3,14 +3,14 @@
     public class StockTickers
     {
         public string T { get; set; }
-        public long c { get; set; }
-        public long h { get; set; }
-        public long l { get; set; }
-        public long n { get; set; }
-        public long o { get; set; }
+        public float c { get; set; }
+        public float h { get; set; }
+        public float l { get; set; }
+        public float n { get; set; }
+        public float o { get; set; }
         public Boolean otc { get; set; }
-        public long t { get; set; }
-        public long v { get; set; }
-        public long vw { get; set; }
+        public float t { get; set; }
+        public float v { get; set; }
+        public float vw { get; set; }
     }
 }
